+++ js_src/_dart_helpers.js
@@ -109,6 +109,14 @@ function _createReactDartComponentClass2(dartInteropStatics, componentStatics, j
     }
 
     static getDerivedStateFromProps(nextProps, prevState) {
+      // Must call checkPropTypes manually because React moved away from using the `prop-types` package.
+      // See: https://github.com/facebook/react/pull/18127
+      // React now uses its own internal cache of errors for PropTypes which broke `PropTypes.resetWarningCache()`.
+      // Solution was to use `PropTypes.checkPropTypes` directly which makes `PropTypes.resetWarningCache()` work.
+      // Solution from: https://github.com/facebook/react/issues/18251#issuecomment-609024557
+      if (process.env.NODE_ENV !== 'production') {
+        React.PropTypes.checkPropTypes(jsConfig.propTypes, nextProps, 'prop', ReactDartComponent2.displayName);
+      }
       let derivedState = dartInteropStatics.handleGetDerivedStateFromProps(componentStatics, nextProps, prevState);
       return typeof derivedState !== 'undefined' ? derivedState : null;
     }
@@ -152,9 +160,6 @@ function _createReactDartComponentClass2(dartInteropStatics, componentStatics, j
     }
     if (jsConfig.defaultProps) {
       ReactDartComponent2.defaultProps = jsConfig.defaultProps;
-    }
-    if (jsConfig.propTypes) {
-      ReactDartComponent2.propTypes = jsConfig.propTypes;
     }
   }